rename scenes file
